+++ src/Sortable.php
@@ -12,11 +12,11 @@ interface Sortable
     /**
      * Let's be nice and provide an ordered scope.
      *
-     * @param \Illuminate\Database\Query\Builder $query
+     * @param \Illuminate\Database\Eloquent\Builder $query
      *
      * @return \Illuminate\Database\Query\Builder
      */
-    public function scopeOrdered(\Illuminate\Database\Query\Builder $query);
+    public function scopeOrdered(\Illuminate\Database\Eloquent\Builder $query);
 
     /**
      * This function reorders the records: the record with the first id in the array